# internal/handler/openapi_embed.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RideMatch API",
    "description": "Backend API for RideMatch — a multi-modal (car, okada, keke, bus) ride/logistics matching platform for Nigeria. Authentication is phone number + OTP based; there is no password field. All responses use a single envelope shape: `{ success, message, data, error }`.",
    "version": "1.0.0",
    "contact": {
      "name": "Joshua Olumoye"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token returned from /auth/otp/verify or /auth/token/refresh. Send as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "APIResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "error": {
            "type": "string"
          }
        }
      },
...
      "CreateTripRequest": {
        "type": "object",
        "required": [
          "vehicle_type",
          "pickup_lat",
          "pickup_lng",
          "destination_lat",
          "destination_lng",
          "offered_price_kobo"
        ],
        "properties": {
          "vehicle_type": {
            "type": "string",
            "enum": [
              "car",
              "okada",
              "keke",
              "bus"
            ],
            "example": "okada"
          },
          "pickup_lat": {
            "type": "number",
            "format": "float"
          },
          "pickup_lng": {
            "type": "number",
            "format": "float"
          },
          "pickup_address": {
            "type": "string"
          },
          "destination_lat": {
            "type": "number",
            "format": "float"
          },
          "destination_lng": {
            "type": "number",
            "format": "float"
          },
          "destination_address": {
            "type": "string"
          },
          "offered_price_kobo": {
            "type": "integer",
            "format": "int64",
            "description": "Passenger's asking price, in kobo (NGN x 100)"
          }
        }
      },
...
      "MakeOfferRequest": {
        "type": "object",
        "required": [
          "price_kobo"
        ],
        "properties": {
          "price_kobo": {
            "type": "integer",
            "format": "int64",
            "description": "Driver's proposed price, in kobo (NGN x 100)"
          }
        }
      },
      "OfferResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "trip_id": {
            "type": "string",
            "format": "uuid"
          },
          "price_kobo": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected",
              "superseded",
              "withdrawn"
            ]
          },
          "driver": {
            "$ref": "#/components/schemas/TripPartyInfo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NearbyOpenTripResponse": {
        "type": "object",
        "properties": {
          "trip_id": {
            "type": "string",
            "format": "uuid"
          },
          "vehicle_type": {
            "type": "string",
            "enum": [
              "car",
              "okada",
              "keke",
              "bus"
            ]
          },
          "pickup_lat": {
            "type": "number",
            "format": "float",
            "description": "Rounded, not exact — see NEARBY_COORDINATE_PRECISION"
          },
          "pickup_lng": {
            "type": "number",
            "format": "float"
          },
          "distance_km": {
            "type": "number",
            "format": "float"
          },
          "offered_price_kobo": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/trips": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Request a trip",
        "description": "Opens a new trip request and instantly notifies nearby online drivers of the matching vehicle type over WebSocket (and indexes it for GET /trips/nearby polling). A passenger may only have one active trip at a time.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trip requested",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TripResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Already have an active trip"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/trips/active": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get my current active trip",
        "description": "Returns the caller's in-progress trip (requested/matched/picked_up), if any.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Active trip fetched",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TripResponse"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No active trip"
          }
        }
      }
    },
    "/trips/nearby": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Find nearby open trip requests",
        "description": "Driver-facing REST fallback for the WebSocket push — open (unmatched) trip requests near a point, for the given vehicle type, nearest first.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "car",
                "okada",
                "keke",
                "bus"
              ]
            }
          },
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby trips fetched",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NearbyOpenTripResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          }
        }
      }
    },
    "/trips/{id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get a trip",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip fetched",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TripResponse"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not a participant"
          },
          "404": {
            "description": "Trip not found"
          }
        }
      }
    },
    "/trips/{id}/offers": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Make a price offer on a trip",
        "description": "A driver proposes a price for an open trip request. Placing a new offer while one is already pending from the same driver updates that offer's price rather than creating a duplicate. Requires the driver to be online, verified, and subscribed.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeOfferRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Offer sent",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OfferResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Not eligible to make offers"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "List pending offers on a trip",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offers fetched",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OfferResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not the trip owner"
          }
        }
      }
    },
    "/trips/{id}/offers/{offerId}/accept": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Accept a driver's offer",
        "description": "Matches the trip to the offering driver and generates a one-time pickup PIN, returned only in this response. Every other offering driver is notified the trip has closed.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver matched",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AcceptOfferResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Trip or offer no longer open"
          },
          "403": {
            "description": "Not the trip owner"
          }
        }
      }
    },
    "/trips/{id}/offers/{offerId}/reject": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Reject a driver's offer",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offer rejected"
          },
          "403": {
            "description": "Not the trip owner"
          }
        }
      }
    },
    "/trips/{id}/confirm-pickup": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Confirm pickup with the passenger's PIN",
        "description": "The driver enters the 4-digit PIN the passenger shows them in person, proving the pickup actually happened.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPickupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pickup confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TripResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Incorrect PIN"
          },
          "403": {
            "description": "Not the matched driver"
          }
        }
      }
    },
    "/trips/{id}/complete": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Mark a trip complete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip completed",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TripResponse"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not the matched driver"
          }
        }
      }
    },
    "/trips/{id}/cancel": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Cancel a trip",
        "description": "Either the passenger or the matched driver may cancel, but only before pickup is confirmed.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelTripRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trip cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TripResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Trip can no longer be cancelled"
          },
          "403": {
            "description": "Not a participant"
          }
        }
      }
    },
...
  }
}